// src/model/dto/people.dto.ts
export type PeopleDto = {
    codigo: string
    cumpleanios: string
    color_ojo: string
    peliculas: string[]
    genero: string
    cabello_color: string
    talla: string
    mundo: string
    peso: string
    nombre: string
    piel_color: string
    creado: string
    editado: string
    especies: string[]
    naves: string[]
    url: string
    vehiculos: string[]
}